Update main.py
I have edited this main.py for Tech_Growth_Predictor. This new update should improve a little better one the html tags identification for most of the data is stored in multiple of identical tags with identical class or id in a node. 
Main goal is to extract a list of information in each nodes such as for <td> tag will yield strings that include company names, issuer name and class of stock; <span> will yield the numerical values of data such as monetary values, shares number, percentage of ownership. Then after having each tag extracted, information will be listed and indexed in a list and only retain the relevant information.
The biggest issue is the website 13F.info itself already has an export to csv function and it seems to be blocking any attempt to extracting its database.
We are attempting to extract information through python and bs4 in order to create a database that is cleaner and extract it straight to sqlite3. 
As of now, I have included a function to export the csv file to sqlite3 just in case it is not possible to scrape through python.